refactor(ui): rename RiPopover files and constants to follow naming conventions
