+++ redisinsight/ui/src/components/base/popover/index.ts
@@ -1,0 +1,2 @@
+export * from './RiPopover'
+export * from './RiPopover.types'
