--- redisinsight/ui/src/components/base/popover/index.tsx
@@ -1,2 +1,0 @@
-export * from './RiPopover'
-export * from './types'
